fix: "방 API 구현"
업로드 시 파일이름에 rooms_id 추가
장비 미등록 버그 수정
테스트 코드 수정

Related to: #6

+++ src/main/java/com/example/airdns/domain/room/dto/RoomsRequestDto.java
@@ -20,7 +20,7 @@ public class RoomsRequestDto {
         private String address;
         private Integer size;
         private String desc;
-        private List<Integer> equipment;
+        private List<Long> equipment;
         private List<String> imageUrl;
     }
